Fix integers enclosed in quotes
TL;DR: No backwards-incompatible API change, no new functionality

Note that the `validate` and `coerce` functions still accept and work
well with duration integers written as strings. This is in line withi
"be liberal in what you accept" and the JavaScript actual good
practices. We want to keep that behavior I think.

So, as it is, this commit is just about documentation and messages
displayed to the user, not about code logic. Everything that worked
before will still work after.