Strict Validation Mode Added
1. Added a Strict Validation mode. If set to true, any properties
specified in config files that are not declared in the schema will
result in errors. This is to ensure that the schema and the config
files are in sync. This brings convict further in line with the concept
of a “Schema”. By default the strict mode is set to false.
2. Added test cases for strict mode