Make schema objects actually useful
The schema nodes were being created by passing the relevant parts of the
schema definition into Object.create(), which creates a new object using
its argument as a prototype. Unfortunately, JSON.stringify() only
includes an object's own properties, i.e. the properties that are not
part of its prototype. As a side effect of this, getSchemaString(),
which returns the stringified schema object, and getSchema(), which
rehydrates that and returns the resulting object, did not include any of
the actual attributes of the properties defined in the schema. By
changing from Object.create() to cloneDeep(), this issue is avoided.

Fixes #161.