Make warnings more visible by coloring them
Warnings are currently easy to miss because they are visually indistinguishable
from other text. It would be better if we write 'Warning:' in bold
and color it in yellow, similar to deprecation warnings from the depd
package.

* If `process.stdout` is a terminal emulator: Use ANSI escape codes
to switch the text weight and color to bold and yellow, write 'Warning:' and
switch text style back to normal.
* Otherwise: Write the string 'Warning:' as usual.

ANSI escape codes are described here:
https://en.wikipedia.org/wiki/ANSI_escape_code